Break up `smithy.protocols.parsers` package
Also add a comment to clarify string argument in `XmlBindingTraitParserGenerator`
