+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/AwsQueryProtocolGenerator.kt
@@ -21,10 +21,10 @@ import software.amazon.smithy.rust.codegen.smithy.generators.StructureGenerator
 import software.amazon.smithy.rust.codegen.smithy.generators.builderSymbol
 import software.amazon.smithy.rust.codegen.smithy.generators.error.errorSymbol
 import software.amazon.smithy.rust.codegen.smithy.generators.operationBuildError
-import software.amazon.smithy.rust.codegen.smithy.protocols.parsers.AwsQueryParserGenerator
-import software.amazon.smithy.rust.codegen.smithy.protocols.parsers.AwsQuerySerializerGenerator
-import software.amazon.smithy.rust.codegen.smithy.protocols.parsers.StructuredDataParserGenerator
-import software.amazon.smithy.rust.codegen.smithy.protocols.parsers.StructuredDataSerializerGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.parse.AwsQueryParserGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.parse.StructuredDataParserGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.serialize.AwsQuerySerializerGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.serialize.StructuredDataSerializerGenerator
 import software.amazon.smithy.rust.codegen.smithy.transformers.OperationNormalizer
 import software.amazon.smithy.rust.codegen.smithy.transformers.RemoveEventStreamOperations
 import software.amazon.smithy.rust.codegen.util.dq
