+++ codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/JsonSerializerGeneratorTest.kt
@@ -1,18 +1,13 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0.
- */
-
-package software.amazon.smithy.rust.codegen.smithy.protocols.parsers
+package software.amazon.smithy.rust.codegen.smithy.protocols.parse
 
 import org.junit.jupiter.api.Test
 import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.model.shapes.StringShape
 import software.amazon.smithy.model.shapes.StructureShape
 import software.amazon.smithy.rust.codegen.rustlang.RustModule
-import software.amazon.smithy.rust.codegen.rustlang.rust
 import software.amazon.smithy.rust.codegen.smithy.generators.EnumGenerator
 import software.amazon.smithy.rust.codegen.smithy.generators.UnionGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.serialize.JsonSerializerGenerator
 import software.amazon.smithy.rust.codegen.smithy.transformers.OperationNormalizer
 import software.amazon.smithy.rust.codegen.smithy.transformers.RecursiveShapeBoxer
 import software.amazon.smithy.rust.codegen.testutil.TestWorkspace
@@ -26,43 +21,39 @@ import software.amazon.smithy.rust.codegen.util.expectTrait
 import software.amazon.smithy.rust.codegen.util.inputShape
 import software.amazon.smithy.rust.codegen.util.lookup
 
-internal class XmlBindingTraitSerializerGeneratorTest {
+class JsonSerializerGeneratorTest {
     private val baseModel = """
         namespace test
-        use aws.protocols#restXml
+        use aws.protocols#restJson1
+
         union Choice {
-            @xmlFlattened
-            @xmlName("Hi")
-            flatMap: MyMap,
-
-            deepMap: MyMap,
-
-            @xmlFlattened
-            flatList: SomeList,
-
-            deepList: SomeList,
-
-            s: String,
-
-            enum: FooEnum,
-
+            blob: Blob,
+            boolean: Boolean,
             date: Timestamp,
-
+            document: Document,
+            enum: FooEnum,
+            int: Integer,
+            list: SomeList,
+            listSparse: SomeSparseList,
+            long: Long,
+            map: MyMap,
+            mapSparse: MySparseMap,
             number: Double,
-
+            s: String,
             top: Top,
-
-            blob: Blob
         }
 
         @enum([{name: "FOO", value: "FOO"}])
         string FooEnum
 
         map MyMap {
-            @xmlName("Name")
             key: String,
+            value: Choice,
+        }
 
-            @xmlName("Setting")
+        @sparse
+        map MySparseMap {
+            key: String,
             value: Choice,
         }
 
@@ -70,19 +61,16 @@ internal class XmlBindingTraitSerializerGeneratorTest {
             member: Choice
         }
 
+        @sparse
+        list SomeSparseList {
+            member: Choice
+        }
 
         structure Top {
             choice: Choice,
-
             field: String,
-
-            @xmlAttribute
             extra: Long,
-
-            @xmlName("prefix:local")
-            renamedWithPrefix: String,
-
-            @xmlFlattened
+            @jsonName("rec")
             recursive: TopList
         }
 
@@ -91,8 +79,10 @@ internal class XmlBindingTraitSerializerGeneratorTest {
         }
 
         structure OpInput {
-            @httpPayload
-            payload: Top
+            @httpHeader("x-test")
+            someHeader: String,
+
+            top: Top
         }
 
         @http(uri: "/top", method: "POST")
@@ -110,25 +100,30 @@ internal class XmlBindingTraitSerializerGeneratorTest {
             )
         )
         val symbolProvider = testSymbolProvider(model)
-        val parserGenerator = XmlBindingTraitSerializerGenerator(testProtocolConfig(model))
-        val operationParser = parserGenerator.payloadSerializer(model.lookup("test#OpInput\$payload"))
+        val parserGenerator = JsonSerializerGenerator(testProtocolConfig(model))
+        val operationGenerator = parserGenerator.operationSerializer(model.lookup("test#Op"))
+        val documentGenerator = parserGenerator.documentSerializer()
 
         val project = TestWorkspace.testProject(testSymbolProvider(model))
         project.lib { writer ->
             writer.unitTest(
                 """
-                 use model::Top;
-                let inp = crate::input::OpInput::builder().payload(
+                use model::Top;
+
+                // Generate the document serializer even though it's not tested directly
+                // ${writer.format(documentGenerator)}
+
+                let input = crate::input::OpInput::builder().top(
                     Top::builder()
                         .field("hello!")
                         .extra(45)
                         .recursive(Top::builder().extra(55).build())
                         .build()
                 ).build().unwrap();
-                let serialized = ${writer.format(operationParser)}(&inp.payload.unwrap()).unwrap();
+                let serialized = ${writer.format(operationGenerator!!)}(&input).unwrap();
                 let output = std::str::from_utf8(serialized.bytes().unwrap()).unwrap();
-                assert_eq!(output, "<Top extra=\"45\"><field>hello!</field><recursive extra=\"55\"></recursive></Top>");
-            """
+                assert_eq!(output, r#"{"top":{"field":"hello!","extra":45,"rec":[{"extra":55}]}}"#);
+                """
             )
         }
         project.withModule(RustModule.default("model", public = true)) {
@@ -141,7 +136,10 @@ internal class XmlBindingTraitSerializerGeneratorTest {
         project.withModule(RustModule.default("input", public = true)) {
             model.lookup<OperationShape>("test#Op").inputShape(model).renderWithModelBuilder(model, symbolProvider, it)
         }
-        println("file:///${project.baseDir}/src/xml_ser.rs")
+        println("file:///${project.baseDir}/src/json_ser.rs")
+        println("file:///${project.baseDir}/src/lib.rs")
+        println("file:///${project.baseDir}/src/model.rs")
+        println("file:///${project.baseDir}/src/operation_ser.rs")
         project.compileAndTest()
     }
 }
